requirements.txt
numpy~=1.24.3
matplotlib~=3.8.0
Pillow~=10.0.0